design pattern updated and some modal upgraded
