--- src/components/Auth/Auth.vue
@@ -1,59 +1,0 @@
-<template>
-  <div class="auth-background">
-    <transition name="slideContainer" mode="out-in" appear>
-      <component :is="getActiveComponent" key="authComponent"></component>
-    </transition>
-  </div>
-</template>
-<script>
-import Login from '../Auth/Login';
-import Signup from '../Auth/Signup';
-import { mapGetters } from 'vuex';
-export default {
-  components: {
-    appLogin: Login,
-    appSignup: Signup
-  },
-  computed: {
-    ...mapGetters(['getActiveComponent'])
-  }
-};
-</script>
-<style scoped>
-
-.slideContainer-enter-active {
-  animation: slide-in 0.3s ease-in-out forwards;
-}
-.slideContainer-leave-active {
-  animation: slide-out 0.3s ease-in-out forwards;
-}
-
-.auth-background {
-  background-image: url('../../assets/img/login.jpg');
-  background-size: auto;
-  background-color: whitesmoke !important;
-  min-height: 100vh;
-  height: auto;
-}
-
-@keyframes slide-in {
-  from {
-    transform: translateX(-1000px);
-    opacity: 0;
-  }
-  to {
-    transform: translateX(0);
-    opacity: 1;
-  }
-}
-@keyframes slide-out {
-  from {
-    transform: translateX(0);
-    opacity: 1;
-  }
-  to {
-    transform: translateX(1000px);
-    opacity: 0;
-  }
-}
-</style>
