+++ src/components/Layout.vue
@@ -43,9 +43,7 @@
     </a-layout-sider>
     <a-layout>
       <a-layout-content :style="{ margin: '24px 16px 0' }">
-        <div :style="{ padding: '24px', background: '#fff', minHeight: '360px'}">
           <slot></slot>
-        </div>
       </a-layout-content>
     </a-layout>
   </a-layout>
