--- src/components/Profile/Profile.vue
@@ -1,14 +1,0 @@
-<template>
-  <div>
-    <app-profile-cover></app-profile-cover>
-  </div>
-</template>
-<script>
-import ProfileCover from '../Profile/ProfileCover/ProfileCover';
-export default {
-  components: {
-    appProfileCover: ProfileCover
-  }
-};
-</script>
-<style scoped></style>
