+++ src/views/Profile.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    <app-profile-cover></app-profile-cover>
+  </div>
+</template>
+<script>
+import ProfileCover from '@/components/ProfileCover';
+export default {
+  components: {
+    appProfileCover: ProfileCover
+  }
+};
+</script>
+<style scoped></style>
